Improve TS typings for `networkInterfaces`
The `networkInterfaces` function normally returns an array, but when
passed the string `default` it will return only a single
`NetworkInterfacesData` object.

This variability makes the current types frustrating to use, because
you'll need to cast the result or check the result at runtime to use it
without the TS compiler complaining. This work is unnecessary, because
the shape of the parameters (available to TS at compile time) determines
the type of the results.

This commit adds three overloads for the most common 0-1 parameter cases
so that the return types won't need casts. The existing declaration is
left as-is (other than making the first parameter required because the
zero-parameter case is already handled with an overload) to handle more
complex cases with callback functions or more than one parameter.